chore: Only job create_iam_role_for_deployment is executed.
Modified the CD workflow to temporarily disable all deployment jobs that depend on the IAM role created in the initial step.
This change ensures that only the first deployment job, responsible for creating or updating the IAM Role for deployment.
This allows for focused application of changes related solely to the IAM role setup.
The subsequent deployment jobs (which utilize the created role) are commented out for now and will be re-enabled in a future commit.